# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trackvance Core",
    "description": "Monolito modular local. SUCCESS es estado técnico; decision conserva el resultado de negocio. Configuraciones y evidence son snapshots versionados.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/connections": {
      "get": {
        "tags": [
          "Conexiones"
        ],
        "summary": "List Connections",
        "operationId": "list_connections_api_v1_connections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:read"
      },
      "post": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Create Connection",
        "operationId": "create_connection_api_v1_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:manage",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/connections/test": {
      "post": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Test Draft",
        "operationId": "test_draft_api_v1_connections_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionTestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:manage",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/connections/{connection_id}": {
      "get": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Connection Detail",
        "operationId": "connection_detail_api_v1_connections__connection_id__get",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:read"
      },
      "patch": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Update Connection",
        "operationId": "update_connection_api_v1_connections__connection_id__patch",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:manage"
      },
      "delete": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Delete Connection",
        "operationId": "delete_connection_api_v1_connections__connection_id__delete",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Version"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:manage"
      }
    },
    "/api/v1/connections/{connection_id}/test": {
      "post": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Test Connection",
        "operationId": "test_connection_api_v1_connections__connection_id__test_post",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:use"
      }
    },
    "/api/v1/connections/{connection_id}/schemas": {
      "get": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Schemas",
        "operationId": "schemas_api_v1_connections__connection_id__schemas_get",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceSchemasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:use"
      }
    },
    "/api/v1/connections/{connection_id}/objects": {
      "get": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Objects",
        "operationId": "objects_api_v1_connections__connection_id__objects_get",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          },
          {
            "name": "schema_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Schema Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceObjectsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:use"
      }
    },
    "/api/v1/connections/{connection_id}/preview": {
      "get": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Preview",
        "operationId": "preview_api_v1_connections__connection_id__preview_get",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          },
          {
            "name": "schema_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Schema Name"
            }
          },
          {
            "name": "object_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Object Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcePreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:use"
      }
    },
    "/api/v1/connections/{connection_id}/datasets": {
      "post": {
        "tags": [
          "Conexiones"
        ],
        "summary": "Register Source Dataset",
        "operationId": "register_source_dataset_api_v1_connections__connection_id__datasets_post",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceDatasetBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "connections:use"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionDeleteResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ConnectionDeleteResponse"
      },
      "ConnectionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConnectionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ConnectionListResponse"
      },
...
      "ConnectionTestResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "SUCCESS",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "tested_at": {
            "type": "string",
            "title": "Tested At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "tested_at"
        ],
        "title": "ConnectionTestResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "LocalSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "trackvance_session",
        "description": "Sesión local HttpOnly. Las mutaciones requieren además X-CSRF-Token."
      }
    }
  }
}